+++ interp/src/interpreter/mod.rs
@@ -1,9 +1,5 @@
-mod interpret_component;
-mod interpret_control;
-mod interpret_group;
 mod steppers;
 mod utils;
 
-pub use interpret_component::interpret_component;
 pub use steppers::{ComponentInterpreter, Interpreter};
 pub use utils::{ConstCell, ConstPort};
